feat: add the category's operations

+++ src/main/java/com/artucorreia/library/domain/entities/Category.java
@@ -4,23 +4,23 @@ import java.util.List;
 import java.util.Objects;
 
 public class Category {
-  private Integer id;
+  private Long id;
   private String name;
   private List<Book> books;
 
   public Category() {}
 
-  public Category(Integer id, String name, List<Book> books) {
+  public Category(Long id, String name, List<Book> books) {
     this.id = id;
     this.name = name;
     this.books = books;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
